Improve code by removing explicit casts to numbers

+++ valiktor-core/src/main/kotlin/org/valiktor/functions/DoubleFunctions.kt
@@ -26,7 +26,7 @@ import org.valiktor.constraints.*
  * @return the same receiver property
  */
 fun <E> Validator<E>.Property<Double?>.isZero(): Validator<E>.Property<Double?> =
-        this.validate(Equals(0.toDouble()), { it == null || it == 0.toDouble() })
+        this.validate(Equals(0.0), { it == null || it == 0.0 })
 
 /**
  * Validates if the [Double] property is not equal to zero
@@ -35,7 +35,7 @@ fun <E> Validator<E>.Property<Double?>.isZero(): Validator<E>.Property<Double?> 
  * @return the same receiver property
  */
 fun <E> Validator<E>.Property<Double?>.isNotZero(): Validator<E>.Property<Double?> =
-        this.validate(NotEquals(0.toDouble()), { it == null || it != 0.toDouble() })
+        this.validate(NotEquals(0.0), { it == null || it != 0.0 })
 
 /**
  * Validates if the [Double] property is equal to one
@@ -44,7 +44,7 @@ fun <E> Validator<E>.Property<Double?>.isNotZero(): Validator<E>.Property<Double
  * @return the same receiver property
  */
 fun <E> Validator<E>.Property<Double?>.isOne(): Validator<E>.Property<Double?> =
-        this.validate(Equals(1.toDouble()), { it == null || it == 1.toDouble() })
+        this.validate(Equals(1.0), { it == null || it == 1.0 })
 
 /**
  * Validates if the [Double] property is not equal to one
@@ -53,7 +53,7 @@ fun <E> Validator<E>.Property<Double?>.isOne(): Validator<E>.Property<Double?> =
  * @return the same receiver property
  */
 fun <E> Validator<E>.Property<Double?>.isNotOne(): Validator<E>.Property<Double?> =
-        this.validate(NotEquals(1.toDouble()), { it == null || it != 1.toDouble() })
+        this.validate(NotEquals(1.0), { it == null || it != 1.0 })
 
 /**
  * Validates if the [Double] property is positive
@@ -62,7 +62,7 @@ fun <E> Validator<E>.Property<Double?>.isNotOne(): Validator<E>.Property<Double?
  * @return the same receiver property
  */
 fun <E> Validator<E>.Property<Double?>.isPositive(): Validator<E>.Property<Double?> =
-        this.validate(Greater(0.toDouble()), { it == null || it > 0.toDouble() })
+        this.validate(Greater(0.0), { it == null || it > 0.0 })
 
 /**
  * Validates if the [Double] property isn't positive
@@ -71,7 +71,7 @@ fun <E> Validator<E>.Property<Double?>.isPositive(): Validator<E>.Property<Doubl
  * @return the same receiver property
  */
 fun <E> Validator<E>.Property<Double?>.isNotPositive(): Validator<E>.Property<Double?> =
-        this.validate(LessOrEqual(0.toDouble()), { it == null || it <= 0.toDouble() })
+        this.validate(LessOrEqual(0.0), { it == null || it <= 0.0 })
 
 /**
  * Validates if the [Double] property is negative
@@ -80,7 +80,7 @@ fun <E> Validator<E>.Property<Double?>.isNotPositive(): Validator<E>.Property<Do
  * @return the same receiver property
  */
 fun <E> Validator<E>.Property<Double?>.isNegative(): Validator<E>.Property<Double?> =
-        this.validate(Less(0.toDouble()), { it == null || it < 0.toDouble() })
+        this.validate(Less(0.0), { it == null || it < 0.0 })
 
 /**
  * Validates if the [Double] property isn't negative
@@ -89,7 +89,7 @@ fun <E> Validator<E>.Property<Double?>.isNegative(): Validator<E>.Property<Doubl
  * @return the same receiver property
  */
 fun <E> Validator<E>.Property<Double?>.isNotNegative(): Validator<E>.Property<Double?> =
-        this.validate(GreaterOrEqual(0.toDouble()), { it == null || it >= 0.toDouble() })
+        this.validate(GreaterOrEqual(0.0), { it == null || it >= 0.0 })
 
 /**
  * Validates if the [Double] property is less than another value
